Rethink how we handle the nonce.
It turns out assuming a random nonce makes less sense than I originally
thought.  Of course we have to ship the nonce across the wire somehow.

This patch reverses the random nonce decision, and adds a new helper
function crypto_secretbox_appendnonce for the case that you want to do
it automagically.

We should probably add a similar thing to crypto_box API, but I don't
need it today.